feat(ButtonDocs): rename DefaultButton with SolidButton and update imports
